+++ webcomponents/remote/src/services/communication/communication.service.tsx
@@ -3,7 +3,12 @@ import * as io from 'socket.io-client';
 import {BehaviorSubject, Observable, Subject} from 'rxjs';
 
 // Types
-import {DeckdeckgoEvent, DeckdeckgoEventSlides, DeckdeckgoEventSlideTo} from '@deckdeckgo/types';
+import {
+  DeckdeckgoEvent,
+  DeckdeckgoEventSlideAction,
+  DeckdeckgoEventSlides,
+  DeckdeckgoEventSlideTo
+} from '@deckdeckgo/types';
 
 const configuration: RTCConfiguration = {
   iceServers: [
@@ -228,7 +233,7 @@ export class CommunicationService {
     this.event.next(data);
   };
 
-  emit(data: DeckdeckgoEvent | DeckdeckgoEventSlides | DeckdeckgoEventSlideTo) {
+  emit(data: DeckdeckgoEvent | DeckdeckgoEventSlides | DeckdeckgoEventSlideTo | DeckdeckgoEventSlideAction) {
     if (this.dataChannelOut) {
       this.dataChannelOut.send(JSON.stringify(data));
     }